requirements.txt
aws-cdk-lib~=2.127.0
constructs~=10.3.0
boto3~=1.34.34
botocore~=1.34.34
pytest==9.0.2
